+++ src/System.Runtime.Extensions/tests/System.Runtime.Extensions.Tests.csproj
@@ -17,15 +17,11 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'netstandard-Windows_NT-Release|AnyCPU'" />
   <ItemGroup Condition="'$(TargetGroup)'=='netstandard' or '$(TargetGroup)'=='netcoreapp'">
     <Compile Include="System\ApplicationIdTests.cs" />
-    <Compile Include="System\IO\PathTests.netstandard.cs" />
-    <Compile Include="System\Convert.netstandard.cs" />
+    <Compile Include="System\Convert.cs" />
     <Compile Include="System\EnvironmentTests.cs" />
-    <Compile Include="System\Math.netstandard.cs" />
     <Compile Include="System\OperatingSystemTests.cs" />
-    <Compile Include="System\StringComparer.netstandard.cs" />
     <Compile Include="System\Runtime\Versioning\VersioningHelperTests.cs" />
     <Compile Include="System\AppDomainTests.cs" />
-    <Compile Include="System\Net\WebUtility.netstandard.cs" />
     <Compile Include="$(CommonTestPath)\System\Runtime\Serialization\Formatters\BinaryFormatterHelpers.cs">
       <Link>Common\System\Runtime\Serialization\Formatters\BinaryFormatterHelpers.cs</Link>
     </Compile>
