+++ src/System.Runtime.Extensions/tests/System/Net/WebUtility.cs
@@ -453,5 +453,27 @@ namespace System.Net.Tests
             }
             return builder.ToString();
         }
+
+        [Theory]
+        [MemberData(nameof(HtmlDecode_TestData))]
+        public static void HtmlDecode_TextWriterOutput(string value, string expected)
+        {
+            if(value == null)
+                expected = string.Empty;
+            StringWriter output = new StringWriter(CultureInfo.InvariantCulture);
+            WebUtility.HtmlDecode(value, output);
+            Assert.Equal(expected, output.ToString());
+        }
+
+        [Theory]
+        [MemberData(nameof(HtmlEncode_TestData))]
+        public static void HtmlEncode_TextWriterOutput(string value, string expected)
+        {
+            if(value == null)
+                expected = string.Empty;
+            StringWriter output = new StringWriter(CultureInfo.InvariantCulture);
+            WebUtility.HtmlEncode(value, output);
+            Assert.Equal(expected, output.ToString());
+        }
     }
 }
